Rename internal directory to "core"
