+++ linkset/src/core/link.ts
