+++ linkset/src/core/linkset.ts
